requirements.txt
Django
psycopg2-binary
python-dotenv
whitenoise
azure-storage-blob
azure-identity
Pillow